// src/global-types/enums/screen-sizes.ts
enum ScreenSizesName {
  ERROR = "error",
  XX_SMALL = "xx-small",
  X_SMALL = "x-small",
  SMALL = "small",
  INTERMEDIATE = "intermediate",
  LARGE = "large",
  X_LARGE = "x-large",
  XX_LARGE = "xx-large"
}

enum ScreenSizesPx {
  XX_SMALL = 320,
  X_SMALL = 360,
  SMALL = 500,
  INTERMEDIATE = 576,
  LARGE = 768,
  X_LARGE = 992,
  XX_LARGE = 1200,
}

export  { ScreenSizesName, ScreenSizesPx };